[Eco-News, DELETE /econews/{econewsId}] The 403 response code is returned

+++ core/src/main/java/greencity/controller/EcoNewsController.java
@@ -7,6 +7,7 @@ import greencity.dto.PageableAdvancedDto;
 import greencity.dto.econews.*;
 import greencity.dto.tag.TagDto;
 import greencity.dto.tag.TagVO;
+import greencity.dto.user.EcoNewsAuthorDto;
 import greencity.dto.user.UserVO;
 import greencity.exception.exceptions.NotFoundException;
 import greencity.service.EcoNewsService;
@@ -248,6 +249,14 @@ public class EcoNewsController {
     @DeleteMapping("/{econewsId}")
     public ResponseEntity<Object> delete(@PathVariable Long econewsId,
                                          @Parameter(hidden = true) @CurrentUser UserVO user) {
+
+        EcoNewsDto ecoNewsDto = ecoNewsService.getById(econewsId);
+        EcoNewsAuthorDto ecoNewsAuthorDto = ecoNewsDto.getAuthor();
+        boolean requestFromCurrentUser = ecoNewsAuthorDto.getId().equals(user.getId());
+        if(!requestFromCurrentUser) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         ecoNewsService.delete(econewsId, user);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
